# openapi.yaml
openapi: 3.0.3
info:
  title: GalaCash API
  description: |
    Backend API for GalaCash - A modern financial management application for class treasurers.

    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the `Authorization` header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Roles
    - `user` - Regular student (siswa)
    - `bendahara` - Class treasurer (bendahara)
  version: 1.0.0
  contact:
    name: GalaCash Team
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://galacash-server-2-66220284668.asia-southeast2.run.app/api
    description: Production server
...
paths:
...
  /cron/generate-bills:
    get:
      tags: [Cash Bills]
      summary: Generate monthly cash bills from Vercel Cron
      description: Protected scheduler endpoint for idempotent monthly bill generation
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer token matching the CRON_SECRET environment variable
          schema:
            type: string
      responses:
        "200":
          description: Bill generation completed
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Cash Bills]
      summary: Generate monthly cash bills manually
      description: Protected manual endpoint for idempotent monthly bill generation
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer token matching the CRON_SECRET environment variable
          schema:
            type: string
      responses:
        "200":
          description: Bill generation completed
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...